+++ packages/portal/stories/components/InlineEdit/index.stories.tsx
@@ -1,11 +1,14 @@
 import * as React from 'react';
-import { withState } from '@dump247/storybook-state';
 import { select, boolean } from '@storybook/addon-knobs';
 import InlineEdit from '@synerise/ds-inline-edit';
 import { action } from '@storybook/addon-actions';
 
+const DEFAULT_VALUE = "Input value";
+
 const stories = {
-  default: withState({ value: '' })(({ store }) => {
+  default: () => {
+    const [value, setValue] = React.useState<string>(DEFAULT_VALUE);
+
     const size = select('Size', ['small', 'normal'], 'normal');
     const widthLimit = boolean('Width limit', false);
     const error = boolean('Error', false);
@@ -16,13 +19,14 @@ const stories = {
         <InlineEdit
           input={{
             name: 'name-of-input',
-            value: store.state.value,
+            value: value,
+            maxLength: 120,
+            placeholder: 'This is placeholder',
             onBlur: action('onBlur'),
+            onChange: event => setValue(event.target.value),
+            onEnterPress: action('onEnterPress'),
           }}
           style={widthLimit ? { maxWidth: 128 } : {}}
-          maxLength={120}
-          onChange={event => store.set({ value: event.target.value })}
-          placeholder={'This is placeholder'}
           size={size}
           error={error}
           disabled={disabled}
@@ -30,7 +34,7 @@ const stories = {
         />
       </div>
     );
-  }),
+  }
 };
 
 export default {