+++ packages/components/icon/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.1](https://github.com/Synerise/synerise-design/compare/@synerise/ds-icon@0.3.0...@synerise/ds-icon@0.3.1) (2019-11-04)
+
+**Note:** Version bump only for package @synerise/ds-icon
+
 # [0.3.0](https://github.com/Synerise/synerise-design/compare/@synerise/ds-icon@0.2.6...@synerise/ds-icon@0.3.0) (2019-10-22)
 
 ### Features