+++ packages/components/status/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.11](https://github.com/Synerise/synerise-design/compare/@synerise/ds-status@0.0.10...@synerise/ds-status@0.0.11) (2019-11-04)
+
+**Note:** Version bump only for package @synerise/ds-status
+
+## [0.0.10](https://github.com/Synerise/synerise-design/compare/@synerise/ds-status@0.0.9...@synerise/ds-status@0.0.10) (2019-11-04)
+
+**Note:** Version bump only for package @synerise/ds-status
+
 ## [0.0.9](https://github.com/Synerise/synerise-design/compare/@synerise/ds-status@0.0.8...@synerise/ds-status@0.0.9) (2019-10-29)
 
 **Note:** Version bump only for package @synerise/ds-status