+++ packages/components/utils/CHANGELOG.md
@@ -3,13 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-## [0.5.1](https://github.com/synerise/synerise-design/compare/@synerise/ds-utils@0.5.0...@synerise/ds-utils@0.5.1) (2019-10-21)
+## [0.5.3](https://github.com/synerise/synerise-design/compare/@synerise/ds-utils@0.5.2...@synerise/ds-utils@0.5.3) (2019-11-04)
 
 **Note:** Version bump only for package @synerise/ds-utils
 
+## [0.5.2](https://github.com/synerise/synerise-design/compare/@synerise/ds-utils@0.5.1...@synerise/ds-utils@0.5.2) (2019-11-04)
 
+**Note:** Version bump only for package @synerise/ds-utils
 
+## [0.5.1](https://github.com/synerise/synerise-design/compare/@synerise/ds-utils@0.5.0...@synerise/ds-utils@0.5.1) (2019-10-21)
 
+**Note:** Version bump only for package @synerise/ds-utils
 
 # [0.5.0](https://github.com/synerise/synerise-design/compare/@synerise/ds-utils@0.4.0...@synerise/ds-utils@0.5.0) (2019-10-15)